refactor(containers):  seperate browser specifics
Strictly moving folders in preparation of simplifying things
