+++ config/paths.js
@@ -78,16 +78,16 @@ module.exports = {
   appSrc: resolveApp('src'),
   appKeys: resolveApp('keys.json'),
   appLocales: resolveApp('src/common/_locales'),
-  appBackground: resolveApp('src/containers/background/background'),
-  appOptions: resolveApp('src/containers/options/options'),
-  appFrameJs: resolveApp('src/containers/save/frame/frame'),
-  appOptionsHTML: resolveApp('src/containers/options/options.html'),
-  appLoginJs: resolveApp('src/containers/auth/login'),
-  appLogoutJs: resolveApp('src/containers/auth/logout'),
-  appSaveJs: resolveApp('src/containers/save/save'),
+  appBackground: resolveApp('src/containers/webkit/background/background'),
+  appOptions: resolveApp('src/containers/webkit/options/options'),
+  appFrameJs: resolveApp('src/containers/webkit/save/frame/frame'),
+  appOptionsHTML: resolveApp('src/containers/webkit/options/options.html'),
+  appLoginJs: resolveApp('src/containers/webkit/auth/login'),
+  appLogoutJs: resolveApp('src/containers/webkit/auth/logout'),
+  appSaveJs: resolveApp('src/containers/webkit/save/save'),
+  appSaveHTML: resolveApp('src/containers/webkit/save/save.html'),
+  appTwitterJs: resolveApp('src/containers/webkit/sites/twitter/twitter'),
   appSafari: resolveApp('src/containers/safari/inject'),
-  appSaveHTML: resolveApp('src/containers/save/save.html'),
-  appTwitterJs: resolveApp('src/containers/sites/twitter/twitter'),
   appManifest: resolveApp('src/manifest.yml'),
   appTsConfig: resolveApp('tsconfig.json'),
   appJsConfig: resolveApp('jsconfig.json'),
