+++ src/containers/webkit/auth/logout.js
