+++ src/store/combineReducers.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux'
-import { setup } from '../containers/background/_setup'
-import { active, tabs } from '../containers/background/_tabs'
-import { saves } from '../containers/save/_save'
-import { tags } from '../containers/save/toolbar/tagging/_tagging'
-import { recommendations } from '../containers/save/recommendations/_recommendations'
-import { survey } from '../containers/save/survey/_survey'
+import { setup } from 'containers/webkit/background/_setup'
+import { active, tabs } from 'containers/webkit/background/_tabs'
+import { saves } from 'containers/webkit/save/_save'
+import { tags } from 'containers/webkit/save/toolbar/tagging/_tagging'
+import { recommendations } from 'containers/webkit/save/recommendations/_recommendations'
+import { survey } from 'containers/webkit/save/survey/_survey'
 
 export default combineReducers({
   setup,
