+++ src/store/combineSagas.js
@@ -2,41 +2,41 @@ import { all, takeLatest } from 'redux-saga/effects'
 import { openPocket } from 'common/helpers'
 import { openOptions } from 'common/interface'
 
-import { wSetup } from '../containers/background/_setup'
-import { wHydrate } from '../containers/background/_setup'
-import { wToggleRecs } from '../containers/background/_setup'
-import { wToggleSite } from '../containers/background/_setup'
-import { wUserLoggedIn } from '../containers/background/_setup'
+import { wSetup } from 'containers/webkit/background/_setup'
+import { wHydrate } from 'containers/webkit/background/_setup'
+import { wToggleRecs } from 'containers/webkit/background/_setup'
+import { wToggleSite } from 'containers/webkit/background/_setup'
+import { wUserLoggedIn } from 'containers/webkit/background/_setup'
 
-import { wTabChanges } from '../containers/background/_tabs'
+import { wTabChanges } from 'containers/webkit/background/_tabs'
 
-import { wSaveTweet } from '../containers/background/_sites'
+import { wSaveTweet } from 'containers/webkit/background/_sites'
 
-import { wAuthCodeRecieved } from '../containers/auth/_auth'
-import { wLogout } from '../containers/auth/_auth'
+import { wAuthCodeRecieved } from 'containers/webkit/auth/_auth'
+import { wLogout } from 'containers/webkit/auth/_auth'
 
-import { wCloseSavePanel } from '../containers/save/_save'
-import { wSavePage } from '../containers/save/_save'
-import { wSaveUrl } from '../containers/save/_save'
-import { wRemoveItem } from '../containers/save/_save'
-import { wArchiveItem } from '../containers/save/_save'
+import { wCloseSavePanel } from 'containers/webkit/save/_save'
+import { wSavePage } from 'containers/webkit/save/_save'
+import { wSaveUrl } from 'containers/webkit/save/_save'
+import { wRemoveItem } from 'containers/webkit/save/_save'
+import { wArchiveItem } from 'containers/webkit/save/_save'
 
-import { wTagChanges } from '../containers/save/toolbar/tagging/_tagging'
-import { wSuggestedTags } from '../containers/save/toolbar/tagging/_tagging'
+import { wTagChanges } from 'containers/webkit/save/toolbar/tagging/_tagging'
+import { wSuggestedTags } from 'containers/webkit/save/toolbar/tagging/_tagging'
 
-import { wRecommendations } from '../containers/save/recommendations/_recommendations'
-import { wSaveRecommendation } from '../containers/save/recommendations/_recommendations'
-import { wOpenRecommendation } from '../containers/save/recommendations/_recommendations'
+import { wRecommendations } from 'containers/webkit/save/recommendations/_recommendations'
+import { wSaveRecommendation } from 'containers/webkit/save/recommendations/_recommendations'
+import { wOpenRecommendation } from 'containers/webkit/save/recommendations/_recommendations'
 
-import { wSpocImpression } from '../containers/save/recommendations/_recommendations'
-import { wSpocView } from '../containers/save/recommendations/_recommendations'
-import { wSpocClick } from '../containers/save/recommendations/_recommendations'
+import { wSpocImpression } from 'containers/webkit/save/recommendations/_recommendations'
+import { wSpocView } from 'containers/webkit/save/recommendations/_recommendations'
+import { wSpocClick } from 'containers/webkit/save/recommendations/_recommendations'
 
-import { wSurvey } from '../containers/save/survey/_survey'
+import { wSurvey } from 'containers/webkit/save/survey/_survey'
 import {
   wSurveyResponse,
   wSurveyCancel
-} from '../containers/save/survey/_survey'
+} from 'containers/webkit/save/survey/_survey'
 
 export function* wOpenPocket() {
   yield takeLatest('OPEN_POCKET', openPocket)
